+++ hubdle-gradle-plugin/main/kotlin/com/javiersc/hubdle/settings/HubdleSettingsPlugin.kt
@@ -70,9 +70,12 @@ constructor(
     }
 
     private fun Settings.createHubdleVersionCatalog() {
-        dependencyResolutionManagement.versionCatalogs.create("hubdle") { builder ->
-            val hubdleCatalogVersion = hubdleSettings.catalogVersion.get()
+        dependencyResolutionManagement.versionCatalogs.register("hubdle") { builder ->
+            val hubdleCatalogVersion: String = hubdleSettings.catalog.version.get()
             builder.from("com.javiersc.hubdle:hubdle-version-catalog:$hubdleCatalogVersion")
+            for ((alias, version) in hubdleSettings.catalog.replaceableVersions.get().entries) {
+                builder.version(alias, version)
+            }
         }
     }